Change Bitmap representation.
The Matrix Ruby's class has a different purpose and 
it doesn't perform updates on the elements of the matrix.

I replaced the representation for a simpler one, keeping the same
interface but making it consistent with the technical document.
Now the Bitmap is stored and accessed in a column-major order.